Modified to use browser API instead and added JSDoc
